Initial commit
Basic working generator

`osview path/to/workspace` will produce `path/to/workspace/index.html`

This is a minimal self-contained view of the actions and their files.

It dynamically loads files when they are clicked on, rather than
preloading them all.

It pre-renders the first 100 lines csv files as a html table. These are
often very large files, even in dev, so can't be rendered fully.

It needs to be served from a webserve, as file:// urls don't work well.